Adds `git` feature with version that adds support for SSH signing (#7)
## Description

This issues relates to #4 where I was able to get local dev containers
to start but alas I ran into some issues. One of the issues was that git
wasn't working. Turns out local dev containers copy over your git config
& in my git config I had SSH signing configured, which as it turns out
Debian bullseye doesn't package a version of git which supports SSH
signing.

I've updated the `devcontainer.json` file, adding two features, one of
which is the latest version of git and the other being the `gh` CLI,
which might come in handy.

We could go further and add features like the AWS CLI, &
Docker-in-Docker, but for the moment this is overkill until we decide
that it's worth teaching this early on.

Lastly, local dev containers won't work for the moment due to a breaking
change in Visual Studio Code v1.75. To get local dev containers working
you will need to revert to 1.74.3 until 1.76 is released in March. There
is also the insiders version of VS Code which includes the fix if you'd
prefer to do that.

Fixes # (issue)

## Type of change

Please select the relevant option.

- [x] Bug fix
- [ ] New feature
- [ ] Breaking change
- [ ] Other

## Checklist

Please review the following checklist and ensure all tasks are completed
before submitting the pull request:

- [x] I have performed a self-review of my own code
- [ ] I have added tests that prove my fix is effective or that my
feature works